% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dres}
\alias{dres}
\title{Dcifer results}
\format{
A three-dimensional array with 52 columns, 52 rows, and 4 matrices.
Dimension names correspond to sample ID's (rows and columns) and types of
results \code{c("estimate", "p_value", "CI_lower", "CI_upper")} (matrices).
}
\usage{
dres
}
\description{
Results of relatedness estimation.
}
\keyword{datasets}
