% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsmp}
\alias{dsmp}
\title{Sample data}
\format{
A list of 52 elements (samples), each element is a list of 87
elements (loci), which are integer vectors (alleles).
}
\usage{
dsmp
}
\description{
Microhaplotype data from two clinics in Mozambique. Samples are sorted by location.
}
\keyword{datasets}
