% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd.R
\name{generateReval}
\alias{generateReval}
\alias{generateRevalList}
\title{Generate a grid of parameter values to evaluate over}
\usage{
generateReval(M, rval = NA, nr = NA)

generateRevalList(Ms, rvals = NA, nrs = NA)
}
\arguments{
\item{M}{an integer.}

\item{rval}{\eqn{{r}} values for the grid. Takes precedence over \code{nr}.}

\item{nr}{an integer. If \code{rval} is not provided, it will be generated
using \code{0}, \code{1}, and \code{nr - 1} values between them.}

\item{Ms}{an integer vector.}

\item{rvals}{a list of the length \code{max(Mv)}. Can also be a vector like
\code{rval}; in that case the same vector will be used for all the values
of \code{Mv}.}

\item{nrs}{an integer vector of the length \code{max(Mv)}.}
}
\value{
A a matrix with \code{M} rows and \code{nr + 1} or
\code{length(rval)} columns.

A list of the length \code{max(Mv)} with elements corresponding to
the values of \code{Mv}. Each element is a matrix with \code{M} rows and
\code{nr + 1} or \code{length(rval)} columns.
}
\description{
Generate a grid of parameter values to evaluate over

Generate a grid of parameter values for multiple values of \code{M}
}
