% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{getCOI}
\alias{getCOI}
\title{Calculate COI}
\usage{
getCOI(dsmp, lrank = 2)
}
\arguments{
\item{dsmp}{a list with each element corresponding to one sample.}

\item{lrank}{the rank of the locus that will determine a sample's COI (loci
are ranked by the number of detected alleles).}
}
\value{
a vector with estimated COI for each sample.
}
\description{
Calculates complexity of infection for a list of samples, using the number of
detected alleles.
}
\examples{
coi <- getCOI(dsmp, lrank = 2)
}
