% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd.R
\name{ibdEstM}
\alias{ibdEstM}
\title{Estimate Relatedness and Number of Related Strains}
\usage{
ibdEstM(
  pair,
  coi,
  afreq,
  nmmax = 6,
  nrs = c(1000, 100, 32, 16, 12, 10),
  rval = NULL,
  revals = NULL,
  logrs = NULL,
  equalr = FALSE,
  out = "mle",
  alpha = 0.05,
  freqlog = FALSE,
  nloc = NULL,
  upcoi = TRUE,
  tol = 1e-09
)
}
\arguments{
\item{pair}{a list of length two containing two samples.}

\item{coi}{a vector indicating complexity of infection for each sample.}

\item{afreq}{a list of allele frequencies. Each element of the list
corresponds to a locus.}

\item{nmmax}{maximum number of related pairs of strains to evaluate over.}

\item{nrs}{an integer vector of values for grid resolution, one value for
each \code{M} (or a single value when \code{equalr = TRUE}).}

\item{rval}{\eqn{{r}} values for the grid or for evaluation when
\code{equalr} is \code{TRUE}. If \code{NULL}, will be evenly spaced between
0 and 1 and interval \eqn{1/nr}.}

\item{revals}{a list where each element is an \code{reval} matrix for a
corresponding \code{M}.}

\item{logrs}{a list where each element is a \code{logr} list for a
corresponding \code{M}.}

\item{equalr}{a logical value. If \code{TRUE}, the same values of \eqn{r} are
assumed for all \code{nm} pairs of related strains.}

\item{out}{a character string for the type of results to be returned. If
\code{"mle"}, an estimate is returned, if \code{"llik"} - a vector of
log-likelihood for each combination.}

\item{alpha}{significance level.}

\item{freqlog}{a logical value indicating if \code{afreq} is on the log
scale.}

\item{nloc}{the number of loci.}

\item{upcoi}{method to handle cases when \eqn{COI < |Ux|} (see @details).}

\item{tol}{tolerance value to decide if a parameter estimate is close enough
to zero.}
}
\value{
\itemize{
\item If \code{out = "mle"}: a vector containing estimated \eqn{{r}}; the
length of the vector is equal to estimated number of related pairs of
strains between the two samples.
\item If \code{out = "llike"}, a vector of log-likelihood values for each
evaluated combination.
\item If \code{out = "all"}, a list containing an estimate, log-likelihood,
maximum log-likelihood, and \eqn{{r}} values corresponding to the
acceptance region determined by the significance level.
}
}
\description{
Estimates multiple relatedness parameters when the number $M$ of
related pairs of strains between two samples is unknown.
}
\seealso{
\code{\link{ibdPair}} for genetic relatedness between two samples
with an option of returning log-likelihood and \code{\link{ibdDat}} for
processing multi-sample data
}
