% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd.R
\name{ibdPair2}
\alias{ibdPair2}
\title{Genetic distance for data with genotyping errors}
\usage{
ibdPair2(
  pair,
  coi,
  afreq,
  nr = 100,
  nm = min(coi),
  rval = NULL,
  reval = NULL,
  logr = NULL,
  equalr = FALSE,
  out = "mle"
)
}
\arguments{
\item{pair}{a list of length two containing two samples.}

\item{coi}{a vector indicating complexity of infection for each sample.}

\item{afreq}{a list of allele frequencies. Each element of the list
corresponds to a locus.}

\item{nr}{an integer value for the resolution of the grid (\eqn{nr - 1}
values between 0 and 1), over which the likelihood will be calculated.
Ignored if non-null \code{reval} is provided.}

\item{nm}{the number of related pairs of strains.}

\item{rval}{\eqn{{r}} values for the grid or for evaluation when
\code{equalr} is \code{TRUE}. If \code{NULL}, will be evenly spaced between
0 and 1 and interval \eqn{1/nr}.}

\item{reval}{the grid of \eqn{{r}} combinations, over which the likelihood
will be calculated. A matrix where each column represents a single
combination.}

\item{logr}{a list as returned by \code{logReval} with logs of \code{reval}
and other quantities.}

\item{equalr}{a logical value. If \code{TRUE}, the same values of \eqn{r} are
assumed for all \code{nm} pairs of related strains.}

\item{out}{a character string for the type of results to be returned. If
\code{"mle"}, an estimate is returned, if \code{"llik"} - a vector of
log-likelihood for each combination.}
}
\value{
\itemize{
\item If \code{out = "mle"}: a vector of length 1 if \code{equalr = TRUE}
or of length \code{nm} otherwise, containing estimated \eqn{{r}} (or vector
/matrix if not just first).
\item If \code{out = "llike"}, a vector of log-likelihood values for each
evaluated combination.
\item If \code{out = "all"}, a list containing an estimate, log-likelihood,
maximum log-likelihood, \eqn{{r}} values corresponding to the acceptance
region determined by the significance level, and the size of that region
(as a proportion of all evaluated).
}
}
\description{
Distance when estimated COI is less than |Ux|: likelihoods for
all subsets of size = COI of Ux are calculated; geometric mean taken.
}
