% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probUxUy.R, R/probUxUyOld.R
\name{logReval}
\alias{logReval}
\title{Calculate logs for reval and associated quantities}
\usage{
logReval(reval, nm = NULL, neval = NULL)

logReval(reval, nm = NULL, neval = NULL)
}
\arguments{
\item{reval}{a grid of \eqn{{r}} combinations, over which the likelihood
will be calculated. A matrix where each column represents a single
combination.}

\item{nm}{the number of related pairs of strains.}

\item{neval}{the number of relatedness values/combinations to evaluate over.}
}
\value{
a list of length \eqn{5} that contains \eqn{log(r)}, \eqn{log(1 - r)}
(matrices of the same dimensions as \code{reval}), the number of \eqn{r =
  1} for each column of \code{reval}, the number of \eqn{0 < r < 1} for each
column, and the sum of \eqn{log(1 - r)} for each column.

a list of length \eqn{5} that contains \eqn{log(r)}, \eqn{log(1 - r)}
(matrices of the same dimensions as \code{reval}), the number of \eqn{r =
  1} for each column of \code{reval}, the number of \eqn{0 < r < 1} for each
column, and the sum of \eqn{log(1 - r)} for each column.
}
\description{
Calculate logs for reval and associated quantities

Calculate logs for reval and associated quantities
}
\details{
Use \code{nm = 1} if \code{equalr = TRUE}.

Use \code{nm = 1} if \code{equalr = TRUE}.
}
