% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probUxUy.R
\name{probUxUyM1}
\alias{probUxUyM1}
\title{Calculate likelihood for a pair
\ifelse{html}{\out{U<sub>x</sub>}}{\eqn{U_x}},
\ifelse{html}{\out{U<sub>y</sub>}}{\eqn{U_y}} when
\ifelse{html}{\out{M = 1}}{\eqn{M = 1}}}
\usage{
probUxUyM1(Ux, Uy, nx, ny, probs, reval, neval, logj, factj)
}
\description{
Calculate likelihood for a pair
\ifelse{html}{\out{U<sub>x</sub>}}{\eqn{U_x}},
\ifelse{html}{\out{U<sub>y</sub>}}{\eqn{U_y}} when
\ifelse{html}{\out{M = 1}}{\eqn{M = 1}}
}
