% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probUxUyOld.R
\name{probUxUyOld}
\alias{probUxUyOld}
\title{Calculate likelihood for a pair
\ifelse{html}{\out{U<sub>x</sub>}}{\eqn{U_x}},
\ifelse{html}{\out{U<sub>y</sub>}}{\eqn{U_y}}}
\usage{
probUxUyOld(Ux, Uy, nx, ny, probs, logr, nm, neval, equalr = FALSE)
}
\arguments{
\item{Ux, Uy}{sets of unique alleles for two samples at a given locus. Vectors
of indices corresponding to ordered probabilities in \code{probs}.}

\item{nx, ny}{complexity of infection for two samples. Vectors of length 1.}

\item{probs}{a vector of population allele frequencies (on a log scale) at a
given locus. It is not checked if frequencies on a regular scale sum to
\eqn{1}.}

\item{logr}{a list of length \eqn{5} as returned by \code{logReval()}.}

\item{nm}{the number of related pairs of strains.}

\item{neval}{the number of relatedness values/combinations to evaluate over.}

\item{equalr}{a logical value. If \code{TRUE}, the same values of \eqn{r} are
assumed for all \code{nm} pairs of related strains.}
}
\value{
A vector of log-likelihood values for each \eqn{{r}}, i.e. for each
column of \code{reval} matrix (or each element of \code{rval} vector if
\code{equalr} is \code{TRUE}).
}
\description{
Calculates log-likelihood for a pair of samples at a single locus.
}
