% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readDat}
\alias{readDat}
\title{Read and Reformat Data}
\usage{
readDat(
  sfile,
  afile = NULL,
  svar = "SampleID",
  lvar = "MarkerID",
  avar = "Haplotype",
  fvar = "frequency",
  ...
)
}
\arguments{
\item{sfile, afile}{name of the files containing sample data and, optionally,
population allele frequencies.}

\item{svar, lvar, avar, fvar}{variable names for sample ID, marker/locus,
allele/haplotype, and population allele frequency.}

\item{...}{additional arguments for \code{read.csv()}.}
}
\description{
Read and Reformat Data
}
