% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rescaleFreq}
\alias{rescaleFreq}
\title{Rescale Allele Frequency}
\usage{
rescaleFreq(afreq, dsmp = NULL, rmrare = FALSE, lbound = 0.01)
}
\arguments{
\item{afreq}{a list with population allele frequencies.}

\item{dsmp}{a list of lists with samples as top level elements.}

\item{rmrare}{a logical value. If \code{TRUE}, rare alleles will be removed from both \code{afreq} and \code{dsmp}; if \code{FALSE}, their frequencies will be increased to \code{lbound}.}

\item{lbound}{a bound for considering an allele "rare".}
}
\value{
a list with updated and rescaled \code{afreq} and \code{dsmp} (updated if \code{rmrare = TRUE}).
}
\description{
Handles small values of estimated population allele frequencies to protect against potential false positives driving the result.
}
