% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revals}
\alias{revals}
\title{Parameter grid}
\format{
A list of length 5, where each element corresponds to a single value
of M and is a matrix with M rows. Each column of a matrix is a
\ifelse{html}{\out{r<sub>1</sub>}}{\eqn{r_1}} = ... =
\ifelse{html}{\out{r<sub>M</sub>}}{\eqn{r_M}}) combination.
}
\usage{
revals
}
\description{
Precalculated parameter grids for a range of values of M (from 1 to 5).
}
\keyword{datasets}
