% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{generateReval}
\alias{generateReval}
\title{Generate a grid of parameter values to evaluate over}
\usage{
generateReval(M, rval = NULL, nr = NULL)
}
\arguments{
\item{M}{an integer.}

\item{rval}{\eqn{{r}} values for the grid. Takes precedence over \code{nr}.}

\item{nr}{an integer. If \code{rval} is not provided, it will be generated
using \code{0}, \code{1}, and \code{nr - 1} values between them.}
}
\value{
A a matrix with \code{M} rows and \code{nr + 1} or
\code{length(rval)} columns.
}
\description{
Generate a grid of parameter values to evaluate over
}
