% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{generateReval}
\alias{generateReval}
\title{Grid of Parameter Values}
\usage{
generateReval(M, rval = NULL, nr = NULL)
}
\arguments{
\item{M}{an integer.}

\item{rval}{a numeric vector with parameter values for the grid. If not
provided, it will be generated by dividing \ifelse{html}{\out{[0,
  1]}}{\eqn{[0, 1]}} into \code{nr} equal intervals.}

\item{nr}{an integer. Ignored if \code{rval} is provided.}
}
\value{
A a matrix with \code{M} rows and \code{nr + 1} or
\code{length(rval)} columns.
}
\description{
Generates a grid of parameter values to evaluate over.
}
\examples{
reval <- generateReval(M = 2, nr = 1e2)
dim(reval)

}
