% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd.R
\name{ibdDat}
\alias{ibdDat}
\title{Pairwise Genetic Distance}
\usage{
ibdDat(
  dsmp,
  coi,
  afreq,
  nr = 100,
  rval = NULL,
  reval = NULL,
  equalr = FALSE,
  out = "all",
  rnull = 0,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{dsmp}{a list containing sample genotypes.}

\item{coi}{a vector indicating complexity of infection for each sample.}

\item{afreq}{a list of allele frequencies. Each element of the list
corresponds to a locus.}

\item{nr}{an integer value for the resolution of the grid (\eqn{nr - 1}
values between 0 and 1), over which the likelihood will be calculated.
Ignored if non-null \code{reval} is provided.}

\item{rval}{\eqn{{r}} values for the grid or for evaluation when
\code{equalr} is \code{TRUE}. If \code{NULL}, will be evenly spaced between
0 and 1 and interval \eqn{1/nr}.}

\item{reval}{the grid of \eqn{{r}} combinations, over which the likelihood
will be calculated. A matrix where each column represents a single
combination.}

\item{equalr}{a logical value. If \code{TRUE}, the same values of \eqn{r} are
assumed for all \code{nm} pairs of related strains.}

\item{out}{a character string for the type of results to be returned. If
\code{"mle"}, an estimate is returned, if \code{"llik"} - a vector of
log-likelihood for each combination.}

\item{rnull}{null value for the relatedness parameter.}

\item{alpha}{significance level.}

\item{...}{additional arguments for \code{ibdPair}.}
}
\value{
A relatedness lower triangular matrix or 3-dimensional array, which
in addition to MLE provides confidence intervals and a p-value for a
specified null.
}
\description{
Pairwise estimation of a relatedness parameter for polyclonal
multiallelic samples.
}
\details{
To be added
}
\seealso{
\code{\link{ibdPair}} for genetic relatedness between two samples
with an option of returning log-likelihood.
}
